+++ SourceCode/Services/ITimeProvider.cs
@@ -3,10 +3,18 @@
     public interface ITimeProvider
     {
         DateTimeOffset Now { get; }
+        DateTime LocalTime { get; }
     }
 
     public class SystemTimeProvider : ITimeProvider
     {
+        public SystemTimeProvider(TimeZoneInfo? timeZoneInfo = null)
+        {
+            TimeZoneInfo = timeZoneInfo ?? TimeZoneInfo.FindSystemTimeZoneById("W. Europe Standard Time");
+        }
         public DateTimeOffset Now => DateTimeOffset.Now;
+        public DateTime LocalTime => TimeZoneInfo.ConvertTimeFromUtc(Now.UtcDateTime, TimeZoneInfo);
+
+        private readonly TimeZoneInfo TimeZoneInfo;
     }
 }